Add BDF to supported solvers
