+++ hydrogym/firedrake/__init__.py
@@ -2,7 +2,7 @@ from hydrogym.core import FlowEnv
 
 from .actuator import DampedActuator
 from .flow import FlowConfig, ScaledDirichletBC
-from .solver import IPCS, NewtonSolver, integrate
+from .solvers import IPCS, NewtonSolver, SemiImplicitBDF, integrate
 from .utils import io, is_rank_zero, linalg, modeling, print
 
 from .envs import Cylinder, Pinball, Cavity, Step  # isort:skip
