+++ hydrogym/firedrake/solvers/__init__.py
@@ -1,0 +1,11 @@
+from .base import NewtonSolver
+from .bdf_ext import SemiImplicitBDF
+from .integrate import integrate
+from .ipcs import IPCS
+
+__all__ = [
+    "NewtonSolver",
+    "IPCS",
+    "SemiImplicitBDF",
+    "integrate",
+]
